Add Cinder Support for OpenStack Plugin
In this version, we mainly implemented the feature of cinder support for
OpenStack Plugin. We finished this by:
1. adding VOLUME_NUMBER and VOLUME_SIZE parameters in the template.
2. while VOLUME_NUMBER and VOLUME_SIZE are both > 0, in allocate method, we will
  allocate and attch cinder volumes to the instance by VolumeApi and
  VolumeAttachmentApi.
3. while VOLUME_NUMBER and VOLUME_SIZE are both > 0, in delete method, we will
  delete the volumes after the instances are deleted (so that we do not need
  to detach the volume).
4. we add some validating codes and retry codes for above code rubostness.
5. we modified some mix-using of spaces and tab to make the codes seem clean.
  We changed all space indents to tab indents.
6. we changed all direct reference to novaApi and cinderApi to call getNovaApi()
  and getCinderApi() to get the APIs.

Other than the above changes, we also change some code to improve the
rubostness of the other part. E.g., we add some validating and retry codes
for floating IP creating and associating, which was proved to be able to avoid
some unexpected issues like no IP allocated or IP not associated properly.